Add seed configuration for tournament formats: Introduce new seed commands for four and five group formats in package.json. Create a shared seeding module to manage tournament configurations and random player generation, streamlining the database seeding process for different tournament formats.
